# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: 食堂饭卡管理系统 API
  description: |
    食堂饭卡管理系统后端接口契约，前后端严格遵照此文档开发。

    **金额单位**：统一为分（int64）。
    **卡号（cardNo）**：16 位数字字符串，独立于数据库自增 ID，所有接口路径参数均使用 cardNo。
    **押金**：固定 2000 分（20 元），系统常量，不由前端传入。
  version: 2.0.0
servers:
  - url: http://localhost:8080
    description: 主后端（开发环境）
  - url: http://localhost:9090
    description: 学籍验证 Mock 服务（开发环境）
...
paths:
...
  /api/validate-student:
    get:
      tags: [validation]
      summary: 校验证件号
      description: |
        发卡前调用，后端转发至学籍验证 Mock 服务，验证证件号是否为本校学生或教职工。
        验证通过后返回姓名和人员类型，前端展示给操作员二次确认后再调发卡接口。
      operationId: validateStudent
      parameters:
        - name: idNumber
          in: query
          required: true
          schema:
            type: string
            pattern: '^\d{12}$'
          description: 证件号，12 位数字
      responses:
        "200":
          description: 证件号有效
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudentValidationResult"
        "404":
          description: 证件号不存在（非本校人员）
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                code: STUDENT_NOT_FOUND
                message: 证件号不存在，非本校学生/教职工
        "502":
          description: 学籍验证服务不可用
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                code: STUDENT_SERVICE_ERROR
                message: 学籍验证服务暂时不可用

  # ==================== 发卡 ====================
  /api/cards:
    post:
      tags: [cards]
      summary: 发卡
      description: |
        操作员二次确认后调用此接口发卡。后端会再次调用学籍验证服务校验证件号。

        业务规则：
        - 证件号必须在学籍验证服务中存在
        - 同一证件号只能持有一张 active 卡，已有 active 卡时返回 409
        - 若该证件号有 lost 状态的卡，系统自动注销旧卡（退还旧卡押金和余额），再创建新卡
        - 每次发卡都生成新的 16 位数字卡号（card_no），不重用旧卡号
        - 押金固定 2000 分（20 元），由系统写入，请求体无需传入
        - 预存款 >= 0
      operationId: issueCard
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IssueCardRequest"
      responses:
        "200":
          description: 发卡成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IssueCardResponse"
        "400":
          description: 参数校验失败（缺少必填项、预存款 < 0 等）
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: 证件号不存在（非本校人员）
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                code: STUDENT_NOT_FOUND
                message: 证件号不存在，非本校学生/教职工
        "409":
          description: 该证件号已有 active 状态的卡
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                code: CARD_ALREADY_ACTIVE
                message: 该证件号已有正在使用的卡

    get:
      tags: [cards]
      summary: 按证件号查询当前有效卡
      description: |
        按证件号（学号/工号）查询该持卡人当前最新的非 cancelled 卡（active 或 lost）。
        主要用于挂失、注销前的查询展示。
        若只有已注销的历史卡或无任何卡，返回 404。
      operationId: getCardByIDNumber
      parameters:
        - name: idNumber
          in: query
          required: true
          schema:
            type: string
            pattern: '^\d{12}$'
          description: 证件号，12 位数字
      responses:
        "200":
          description: 查询成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CardDetail"
        "404":
          description: 该证件号无有效卡
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                code: CARD_NOT_FOUND
                message: 该证件号无有效卡

  # ==================== 查卡 ====================
  /api/cards/{cardNo}:
    get:
      tags: [cards]
      summary: 按卡号查询卡信息
      description: |
        根据 16 位卡号查询卡片详情，包含持卡人信息。
        无论卡片状态如何（active/lost/cancelled）都会返回，前端根据 status 字段判断。
        用于就餐刷卡和存款时的查询。
      operationId: getCard
      parameters:
        - $ref: "#/components/parameters/CardNo"
      responses:
        "200":
          description: 查询成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CardDetail"
        "404":
          description: 卡号不存在（非本单位所发）
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                code: CARD_NOT_FOUND
                message: 卡号不存在

  # ==================== 存款 ====================
  /api/cards/{cardNo}/deposits:
    post:
      tags: [deposits]
      summary: 存款（充值）
      description: |
        为指定卡号充值。充值成功后返回收据信息（卡号、持卡人姓名、充值金额、充值后余额、充值时间）。

        约束：
        - 充值金额必须 > 0
        - 卡必须为 active 状态
      operationId: deposit
      parameters:
        - $ref: "#/components/parameters/CardNo"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DepositRequest"
      responses:
        "200":
          description: 充值成功，返回收据信息
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DepositResponse"
        "400":
          description: 充值金额无效
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: 卡号不存在
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: 卡不是 active 状态（已挂失或已注销）
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                code: CARD_NOT_ACTIVE
                message: 该卡已挂失，无法充值

  # ==================== 就餐消费 ====================
  /api/cards/{cardNo}/transactions:
    post:
      tags: [transactions]
      summary: 就餐消费（结算扣款）
      description: |
        窗口工作人员确认结算后调用此接口扣款。

        后端执行校验：
        1. 卡号存在
        2. 卡状态不是 cancelled
        3. 卡状态不是 lost
        4. 余额充足

        前端流程：先调用 GET /api/cards/{cardNo} 展示卡信息和余额，再调用此接口结算。
      operationId: createTransaction
      parameters:
        - $ref: "#/components/parameters/CardNo"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TransactionRequest"
      responses:
        "200":
          description: 扣款成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransactionResponse"
        "400":
          description: 消费金额无效
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: 卡号不存在
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                code: CARD_NOT_FOUND
                message: 此卡非本单位所发
        "409":
          description: 业务校验失败（卡已注销、卡已挂失、余额不足）
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              examples:
                cancelled:
                  summary: 卡已注销
                  value:
                    code: CARD_CANCELLED
                    message: 此卡已注销
                lost:
                  summary: 卡已挂失
                  value:
                    code: CARD_LOST
                    message: 此卡已挂失
                insufficient:
                  summary: 余额不足
                  value:
                    code: INSUFFICIENT_BALANCE
                    message: 余额不足

  # ==================== 挂失 ====================
  /api/cards/{cardNo}/loss-report:
    put:
      tags: [loss-report]
      summary: 挂失
      description: |
        将 active 状态的卡挂失（状态变为 lost）。
        挂失后在任何窗口刷卡都会被拦截报警。

        前端流程：先调用 GET /api/cards?idNumber=xxx 按证件号找到卡，展示后再调用此接口。
      operationId: reportLoss
      parameters:
        - $ref: "#/components/parameters/CardNo"
      responses:
        "200":
          description: 挂失成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CardDetail"
        "404":
          description: 卡号不存在
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: 卡不是 active 状态
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                code: CARD_NOT_ACTIVE
                message: 只有正常使用中的卡可以挂失

    delete:
      tags: [loss-report]
      summary: 取消挂失
      description: |
        将 lost 状态的卡恢复为 active。

        前端流程：先调用 GET /api/cards?idNumber=xxx 按证件号找到卡，展示后再调用此接口。
      operationId: cancelLossReport
      parameters:
        - $ref: "#/components/parameters/CardNo"
      responses:
        "200":
          description: 取消挂失成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CardDetail"
        "404":
          description: 卡号不存在
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: 卡不是 lost 状态
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                code: CARD_NOT_LOST
                message: 只有已挂失的卡可以取消挂失

  # ==================== 注销 ====================
  /api/cards/{cardNo}/cancellation:
    post:
      tags: [cancellation]
      summary: 注销
      description: |
        注销卡片，退还押金和余额。

        约束：
        - 卡状态为 active 或 lost 时可注销
        - 已注销的卡不能再注销
        - 注销后余额清零，状态变为 cancelled

        前端流程：先调用 GET /api/cards?idNumber=xxx 按证件号找到卡，展示后再调用此接口。
      operationId: cancelCard
      parameters:
        - $ref: "#/components/parameters/CardNo"
      responses:
        "200":
          description: 注销成功，返回应退金额
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CancellationResponse"
        "404":
          description: 卡号不存在
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: 卡已注销
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                code: CARD_ALREADY_CANCELLED
                message: 该卡已注销

  # ==================== 汇总统计 ====================
...
components:
  parameters:
    CardNo:
      name: cardNo
      in: path
      required: true
      schema:
        type: string
        pattern: '^\d{16}$'
      description: 16 位数字卡号（card_no，非数据库自增 ID）
...
  schemas:
...
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          description: |
            错误码枚举：
            - CARD_NOT_FOUND: 卡号不存在
            - CARD_ALREADY_ACTIVE: 该证件号已有 active 卡
            - CARD_NOT_ACTIVE: 卡不是 active 状态
            - CARD_NOT_LOST: 卡不是 lost 状态
            - CARD_CANCELLED: 卡已注销
            - CARD_LOST: 卡已挂失
            - CARD_ALREADY_CANCELLED: 卡已注销，不能重复注销
            - INSUFFICIENT_BALANCE: 余额不足
            - INVALID_AMOUNT: 金额无效
            - WINDOW_NOT_FOUND: 窗口不存在
            - VALIDATION_ERROR: 通用校验错误
            - STUDENT_NOT_FOUND: 证件号不存在（非本校人员）
            - STUDENT_SERVICE_ERROR: 学籍验证服务不可用
            - INTERNAL_ERROR: 服务端内部错误
        message:
          type: string

    # ---------- 学籍验证 ----------
    StudentValidationResult:
      type: object
      required: [valid, idNumber, name, type]
      properties:
        valid:
          type: boolean
        idNumber:
          type: string
          description: 12 位证件号
        name:
          type: string
          description: 姓名
        type:
          type: string
          enum: [student, staff]
          description: 人员类型（学生/教职工）

    # ---------- 基础实体 ----------
    CardHolder:
      type: object
      required: [id, name, idNumber, createdAt]
      properties:
        id:
          type: integer
          format: int64
          description: 内部自增 ID
        name:
          type: string
        idNumber:
          type: string
          description: 12 位证件号
        createdAt:
          type: string
          format: date-time
    Card:
      type: object
      required: [cardNo, cardHolderId, deposit, balance, status, createdAt, updatedAt]
      properties:
        cardNo:
          type: string
          pattern: '^\d{16}$'
          description: 16 位数字卡号（对外唯一标识）
        cardHolderId:
          type: integer
          format: int64
        deposit:
          type: integer
          format: int64
          description: 押金（分），固定 2000
        balance:
          type: integer
          format: int64
          description: 余额（分）
        status:
          type: string
          enum: [active, lost, cancelled]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CardDetail:
      description: 卡片详情（含持卡人信息）
      allOf:
        - $ref: "#/components/schemas/Card"
        - type: object
          required: [cardHolder]
          properties:
            cardHolder:
              $ref: "#/components/schemas/CardHolder"
...
    IssueCardRequest:
      type: object
      required: [idNumber]
      properties:
        idNumber:
          type: string
          pattern: '^\d{12}$'
          description: 持卡人证件号，12 位数字
        preDeposit:
          type: integer
          format: int64
          minimum: 0
          default: 0
          description: 预存款（分），可以为 0
    IssueCardResponse:
      type: object
      required: [card, cardHolder]
      properties:
        card:
          $ref: "#/components/schemas/Card"
        cardHolder:
          $ref: "#/components/schemas/CardHolder"
        refund:
          description: 旧卡退款信息（仅当证件号有 lost 卡被自动注销时返回）
          nullable: true
          type: object
          required: [oldCardNo, deposit, balance, total]
          properties:
            oldCardNo:
              type: string
              description: 被注销的旧卡号（16 位）
            deposit:
              type: integer
              format: int64
              description: 退还押金（分）
            balance:
              type: integer
              format: int64
              description: 退还余额（分）
            total:
              type: integer
              format: int64
              description: 应退总额（分）= deposit + balance

    # ---------- 存款 ----------
    DepositRequest:
      type: object
      required: [amount]
      properties:
        amount:
          type: integer
          format: int64
          minimum: 1
          description: 充值金额（分），必须大于 0
    DepositResponse:
      description: 充值成功的收据信息
      type: object
      required: [id, cardNo, holderName, amount, newBalance, createdAt]
      properties:
        id:
          type: integer
          format: int64
          description: 存款记录 ID
        cardNo:
          type: string
          description: 16 位卡号
        holderName:
          type: string
        amount:
          type: integer
          format: int64
          description: 充值金额（分）
        newBalance:
          type: integer
          format: int64
          description: 充值后余额（分）
        createdAt:
          type: string
          format: date-time

    # ---------- 就餐消费 ----------
    TransactionRequest:
      type: object
      required: [windowId, amount]
      properties:
        windowId:
          type: integer
          format: int64
        amount:
          type: integer
          format: int64
          minimum: 1
          description: 消费金额（分），必须大于 0
    TransactionResponse:
      type: object
      required: [id, cardNo, windowId, amount, newBalance, createdAt]
      properties:
        id:
          type: integer
          format: int64
        cardNo:
          type: string
          description: 16 位卡号
        windowId:
          type: integer
          format: int64
        amount:
          type: integer
          format: int64
        newBalance:
          type: integer
          format: int64
        createdAt:
          type: string
          format: date-time

    # ---------- 注销 ----------
    CancellationResponse:
      type: object
      required: [card, refund]
      properties:
        card:
          $ref: "#/components/schemas/Card"
        refund:
          type: object
          required: [deposit, balance, total]
          properties:
            deposit:
              type: integer
              format: int64
              description: 退还押金（分）
            balance:
              type: integer
              format: int64
              description: 退还余额（分）
            total:
              type: integer
              format: int64
              description: 应退总额（分）= deposit + balance